updated schema

+++ schema/world_bank_schema.py
@@ -3,9 +3,9 @@ from pyspark.sql.types import StructType,StructField,IntegerType,FloatType,Strin
 data_schema = StructType([
     StructField("Country", StringType(), True),
     StructField("Year", IntegerType(), True),
-    StructField("GDP_USD", IntegerType(), True),
-    StructField("Population",IntegerType(),True),
-    StructField("Life_Expectancy",IntegerType(),True),
+    StructField("GDP_USD", FloatType(), True),
+    StructField("Population",FloatType(),True),
+    StructField("Life_Expectancy",FloatType(),True),
     StructField("Unemployment_Rate",FloatType(),True),
     StructField("CO2_Emission",FloatType(),True),
     StructField("Electricity_Access",FloatType(),True),
